Fix bug that was causing the rbind to create a gigantic output dataset
Instead of a single NA in the legacy dataset for each variable, a huge string of NAs was being created.  I'm not sure why it started happening but I fixed by switching to dplyr::bind_rows() which
doesn't require the columns to match exactly. But it did require some explicit type matching, which I added by coercing the legacy types to the standard types.